chore: revert SESSION_MAX_HOURS to int for type consistency (V9.58)

+++ bbs_bot_v9.py
@@ -87,7 +87,7 @@ class BotConfiguration:
 
     # Operational Safety
     MAX_CONSECUTIVE_RECOVERIES: int = 3
-    SESSION_MAX_HOURS: float = 16
+    SESSION_MAX_HOURS: int = 16
     POLL_MAIN_LOOP: float = 0.1
     POLL_UI_VERIFY: float = 0.1
     POLL_POPUP: float = 0.5
